requirements.txt
argcomplete==2.0.0
asgiref==3.6.0
attrs==23.1.0
Automat==22.10.0
Babel==2.11.0
Beaker==1.10.0
beautifulsoup4==4.12.2
blivet==3.7.1
blivet-gui==2.4.1
Brlapi==0.8.4
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cupshelpers==1.0
customtkinter==5.1.2
darkdetect==0.8.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
distro==1.8.0
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
gpg==1.17.1
gssapi==1.7.3
gunicorn==20.1.0
humanize==3.13.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
koji==1.32.0
langtable==0.0.61
libcomps==0.1.18
lxml==4.9.2
Mako==1.2.3
MarkupSafe==2.1.2
MouseInfo==0.1.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nftables==0.1
numpy==1.24.3
olefile==0.46
opencv-python==4.7.0.72
packaging==23.0
Paste==3.5.2
pexpect==4.8.0
pid==2.2.3
Pillow==9.4.0
ply==3.11
productmd==1.35
progressbar2==3.53.2
protobuf==3.20.3
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycairo==1.23.0
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.17
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
PyGetWindow==0.0.9
pygit2==1.12.0
PyGObject==3.44.1
PyJWT==2.6.0
pykickstart==3.47
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparted==3.12.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python-utils==3.3.3
python3-xlib==0.15
pytweening==1.0.5
pytz==2023.3
pyudev==0.24.0
pyxdg==0.27
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
rest-framework-simplejwt==0.0.2
rpm==4.18.1
rpmautospec==0.3.5
selinux==3.5
sepolicy==3.5
setools==4.4.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.5.1
soupsieve==2.3.2.post1
sqlparse==0.4.4
systemd-python==235
Tempita==0.5.2
tk==0.1.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.12
vidstream==0.0.14
vncdotool==1.1.0
zope.interface==6.0
argcomplete==2.0.0
asgiref==3.6.0
attrs==23.1.0
Automat==22.10.0
Babel==2.11.0
Beaker==1.12.1
beautifulsoup4==4.12.2
blinker==1.6.2
blivet==3.7.1
blivet-gui==2.4.1
bottle==0.12.25
Brlapi==0.8.4
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cupshelpers==1.0
customtkinter==5.1.2
darkdetect==0.8.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
distro==1.8.0
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fedora-third-party==0.10
file-magic==0.4.0
Flask==2.3.2
fros==1.1
gpg==1.17.1
gssapi==1.7.3
gunicorn==20.1.0
humanize==3.13.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
koji==1.32.0
langtable==0.0.62
libcomps==0.1.18
lxml==4.9.2
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.2
MouseInfo==0.1.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nftables==0.1
numpy==1.24.3
olefile==0.46
opencv-python==4.7.0.72
packaging==23.0
Paste==3.5.2
pexpect==4.8.0
pid==2.2.3
Pillow==9.5.0
ply==3.11
productmd==1.35
progressbar2==3.53.2
protobuf==3.20.3
proxy-tools==0.1.0
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycairo==1.23.0
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.17
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
PyGetWindow==0.0.9
pygit2==1.12.0
PyGObject==3.44.1
PyJWT==2.6.0
pykickstart==3.47
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparted==3.12.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python-utils==3.3.3
python3-xlib==0.15
pytweening==1.0.5
pytz==2023.3
pyudev==0.24.0
pywebview==4.1
pyxdg==0.27
PyYAML==6.0
QtPy==2.3.1
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
rest-framework-simplejwt==0.0.2
rpm==4.18.1
rpmautospec==0.3.5
selinux==3.5
sepolicy==3.5
setools==4.4.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.5.1
soupsieve==2.3.2.post1
sqlparse==0.4.4
systemd-python==235
Tempita==0.5.2
tk==0.1.0
tkhtmlview==0.2.0
tkinterhtml==0.7
tkinterweb==3.18.8
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.12
vidstream==0.0.14
vncdotool==1.1.0
Werkzeug==2.3.4
zope.interface==6.0
